Implement support for evaluating call frame unwinding programs
This commit adds support for evaluating `CallFrameInstruction`s, yielding what
the standard calls the "virtual unwind table". The virtual unwind table is
represented here as a streaming iterator over its rows.

To minimize the number of allocations required when repeatedly running many CFI
programs (for example, when unwinding the whole stack), the context in which a
CFI program runs has been split out into its own re-usable type. Furthermore,
because the `DW_CFA_restore` instruction allows a register's rule to be restored
to the value defined after evaluating the CIE's initial instructions but before
evaluating any of the FDE's instructions, we need to keep track of the initial
register rules and if we evaluated them or not yet. To shelter library users
from tracking this state manually, we push it into the type system via
new-types (`InitializedUnwindContext` and `UninitializedUnwindContext`), and
only these new-types are exposed publicly.