line: don't store references in LineNumberRow
This allows users to clone the row and store it in a Vec, for example.

This is an API change only.  It shouldn't actually change the code
that is generated.  Before, next_row() returned a struct containing
two references. Now, it returns a tuple containing two references.

The problem before was that the struct contained a reference to
StateMachineRegisters, but the user couldn't clone that because it is
private.