Clean up and fix `.eh_frame` parsing
Rather than putting methods for *parsing* the parts of CFI entries that change
between `.eh_frame` and `.debug_frame` and various versions in
`_UnwindSectionPrivate`, this makes it so that we return the *encoding* used (if
any). This lets us understand the difference both when parsing and when using
`test-assembler` to write test data.

Additionally, it has fixes for `.eh_frame` parsing where certain things (like
address and segment sizes) do not appear at all in the CIEs.