Use a Vec<(number, rule)> to keep track of register rules
What we were doing was constructing an array where each index I contained the
recovery rule for register I. We usually have very few register rules: only one
or two. The result was that we would allocate more space than we needed and fill
it with undefined register rules. Avoiding that is a nice little speedup, as
shown in the `cargo benchcmp` results below:

```
 name                                                       before ns/iter  after ns/iter  diff ns/iter   diff %
 cfi::eval_longest_fde_instructions_new_ctx_everytime       539             419                    -120  -22.26%
 cfi::eval_longest_fde_instructions_same_ctx                504             396                    -108  -21.43%
 cfi::iterate_entries_evaluate_every_fde                    2,247,613       1,735,106          -512,507  -22.80%
```