Make CFI evaluation signal-safe
This commit modifies CFI parsing and evaluation so that it is signal safe. This
mostly involved the removal of dynamic allocation from using
`std::vec::Vec`. Now, we use `arrayvec` in `std::vec::Vec`'s place. This means
we pre-allocate space for register rules and the CFI evaluation stack, and it is
now more important for CFI structures to be space conscious. With that
motivation, I changed the representation of registers to `u8` rather than `u64`.

Finally, I tried to annotate all `impl` blocks whose methods should be
signal-safe. The hope is that this will serve as documentation for API users (it
shows up in the rustdocs) and guidelines for gimli hackers. In the future, it
would be great to implement a lint that enforces this property.