Update top level docs
Remove "marketing"-y bits from crate docs -- we already have that in the README
and if you're reading docs you don't need to be hit over the head with this
stuff anymore.

But you might like to see the list of example programs, so add that in.

Also add a section enumerating cargo features that can be flipped on or off.