Optimize EndianReader to use ptr/len
This requires `stable_deref_trait::CloneStableDeref` for the bytes. If
needed in future, we could relax that to `stable_deref_trait::StableDeref`
by implementing a `Clone` for `SubRange` that recalculates the `ptr`.

This doesn't show any significant change in the gimli benchmarks, but I
have modified `addr2line` to use `Rc<Cow<[u8]>>`, and it shows about 15%
improvement for all operations in that (both context creation and
lookup).