read: don't use rnglists and loclists header
These headers don't give information that is required for parsing
range lists or location lists, because the encoding of these lists
must match the encoding of the compilation unit.

Additionally, while the standard seems to indicate that each section
has only one of these headers, in practice this is not true.
Furthermore, when there are multiple headers, they do not have to
each have the same encoding, so our parsing was wrong in this case
because we only used the encoding from the first header.