read: some producers use data4/data8 for section offsets
DW_FORM_data4/DW_FORM_data8 should only be used for section offsets
in DWARF 2/3, but some producers use them for DWARF 4 as well.

For most attributes, this isn't ambiguous, so allow them as well.
The exception is DW_AT_data_member_location.

This is a regression introduced in 29a86ed6