Delete phantom Section members
These complicate the API without providing significant benefit.
Deleting these allows consumers to reuse more code between
.eh_frame and .debug_frame.

The main benefit I could see for these was in the tests, where it
ensured that the section type we wrote in the test builder was the
same as the section type we tried to read back. We can retain this
check by adding a `SectionKind` parameter in the tests, and we can
extend this check to handle endianity as well.