Prepare for inclusion into libstd
This commit adds the "standard preamble" in `Cargo.toml` which makes
this crate suitable for being used as a dependency of the standard
library. The intention here is that this is a "fire and forget" snippet
added to `Cargo.toml` where it doesn't really impact anything else in
this crate. It's unlikely to be used too too soon for libstd, but I
figure it wouldn't be bad to get ready!