Remove dependency of `read` on `stable_deref_trait`
This commit adds a new feature to this crate, enabled by default, called
`endian-reader`. The goal here is to remove the need for the `backtrace`
crate to depend on `stable_deref_trait`, and it looks like currently
`addr2line` and friends don't need the only usage of
`stable_deref_trait`, which is the `EndianReader` type.

Another possible solution might be to break up the `read` feature, too,
and have `addr2line` only enable a small subset of the `read` feature
that it actually needs.