Remove Travis CI; Switch to Github Actions
Previous commits set up an initial install of github actions, this commit is
mostly removing Travis CI. Note that it does not completely remove Travis CI,
since our coverage job is not fully ported over yet. We do run the coverage job
in github actions, but don't upload it anywhere, currently. So for now, Travis
CI is still enabled, but only runs the coverage job and uploads the results to
coveralls.io.