Add EntriesRaw::skip_attributes.
This makes it easier for consumers to skip over abbrevs that
they don't need. It might also be slightly faster than parsing
the unneeded attributes and throwing away their values, but
any speed benefits are unproven so far.
The performance of this is dominated by memory accesses: Either
reading the bytes from main memory into the cache, or even paging
the file contents in from disk. The skipping in skip_attributes
only saves memory accesses for fixed-size attributes, but variable-
length attributes are so common that we're never able to skip an
entire cache line.