When making Dwarfs for DWOs, copy over the supplementary file from the parent.
The spec contemplates that split DWARF and supplementary files can be used together
in section B.2. The only tool I know of that generates supplementary files is
dwz, which doesn't have any support for split DWARF yet. But there's no inherent
reason they can't be used together. While it wouldn't make much sense to do this
with unpackaged dwos, it would make a lot of sense to do it with packaged dwps.