Delete From and Index trait impls for EndianSlice
The From implementation for &[u8] can cause type inference problems.
We could change back to only having an Into implementation, but in practice
EndianSlice is never used in a generic context that requires this trait,
and the &[u8] is easily obtained by either the `slice` method or Deref.

Remove the Index implementation while we're at it, since this is
also not useful in practice.