read: add Evaluation::value_result
DWARF expressions can be used to either compute a value or a location
description, depending on context.

Currently a value computation will result in a location description that
contains one `Piece` whose `location` is set to `Location::Address`,
and this address is the computed value.

`Evaluation::value_result` makes it easier to obtain this value,
supports results that are typed values, and checks there were no
operations that only valid for location descriptions.