Report more specific errors for unknown constants
In case someone encounters some DWARF containing a constant value not
known to gimli, let's ensure that the error has the actual constant
available in it.

These values are not visible in `Error::description` or the various
other error-reporting functions which use it, such as the `Display`
implementation. I think it would be good to improve that too, but wanted
to keep this PR relatively small.