read/cfi: add parse_encoded_value
This directly parses an encoded value, instead of relying on
parse_encoded_pointer to skip the pointer handling.
This is preparation for changing parse_encoded_pointer to
return a different address type.

Also add more validation for the FDE count encoding, since it
seems like the right thing to do, but this could be removed again
if it is a problem.