read/cfi: type parameter style changes
This is preparation for an Address type parameter.

Rename some storage type parameters from A to S. This is for consistency
with other existing uses, and frees up A to be used for Address.

Also use where clauses for long type parameter lists, since these
will be even longer with the Address parameter.