write: add `ConvertLine`
This allow reuse of the implementation of `LineProgram::from` as part
of a more complex transformation.

Operation of `LineProgram::from` is mostly unchanged. The one difference
is that it now uses the string form chosen by `LineString::new`
instead of copying the form of the input.

Also relax the sequence constraints on `write::LineProgram`. We now
automatically start sequences if needed. This better matches how the
underlying DWARF instructions operate, and simplifies use of
`LineConvert`.