Relax the sequence constraints on `write::LineProgram`
We now automatically start sequences if needed.

This better matches how the underlying DWARF instructions operate,
and simplifies use of `LineConvert` since we no longer need to
check if a sequence has started.

This also means that `LineConvert::read_row` will work as expected
if there are multiple `DW_LNE_set_address` in a sequence.
`LineConvert::read_sequence` still does not support this.