Add support for reading items from .debug_macinfo (#759)
The section .debug_macinfo was specified in DWARF standards v2, v3 and v4. It was replaced by .debug_macro in DWARF v5.
The content of .debug_macinfo is represented through the enum DebugMacInfoItem.
It has variants for each specified type:
- macro defined
- macro undefined
- file includd
- end of include
- vendor extension
- end of unit

A DW_AT_macro_info attribute points to a list of macros that were defined in the unit.
To handle this, get_macinfo(offset) returns an iterator over the items found starting at the offset.

This change does not implement writing of .debug_macinfo.