Remove #[non_exhaustive] from read::CallFrameInstruction
Using #[non_exhaustive] disallows users of the library from exhaustively
matching the enum. This is nice for semver, but terrible for library
users who are now in the awkward position of having to remember to
update their exhaustive match when updating gimli.

This is less than ideal. I would know, as I now have to deal with it.