write/unit: ignore DW_AT_GNU_locviews when converting attributes (#771)
We don't yet support parsing the .debug_loc entries for
DW_AT_GNU_locviews, and adding support for it is a low priority.
DW_AT_GNU_locviews was designed so that consumers which don't
understand it can ignore it, so this commit does that for the
case of converting attributes in `write::UnitTable::from`.