Add constants obsolete after DWARF 1.1 (#775)
The DWARF 1.1 specification includes some constants which are only
relevant in that spec, and became obsolete in the later revisions of
the spec.

Currently the gimli does not include these constants because since
DWARF 1.1 is a different structure and rarely relevant it does not
support parsing of DWARF 1.1.

However DWARF 1.1 is relevant to the decomp community. Since decomp
tooling wants to depend on gimli anyways for DWARF 2+, it would be
nice if it could also use at least gimli data structures for
DWARF 1.1 rather than having to write something entirely bespoke.
Including the constants unique to DWARF 1.1 helps with that.

This is unlikely to cause much confusion about what is supported since
anyone who runs into DWARF 1.1 in practice is probably already deep
in the weeds.