Add a llvm-dwarfdump inspired --debug-info [OFFSET] argument to dwarfdump. (#776)
This dumps only .debug_info and if the optional offset is provided only
the subtree of DIEs starting at that offset.

This relies on the way parallel_output() outputs each unit to its own individual buffer
that can be discarded partway through. Implementing it for .debug_types isn't possible
at the moment because dump_types() outputs directly to stdout.