Keep the string form of DW_LNCT_LLVM_source on conversion
With the existing conversion from a readable DWARF representation to
one that is writable the DW_LNCT_LLVM_source string form is kept.
When an empty source string is needed, it currently uses
DW_FORM_string and a mismatch with DW_LNCT_LLVM_source results in an
error when serializing.

Instead of always using DW_FORM_string when emitting empty source
entries, use the same form as DW_LNCT_LLVM_source has.  This is only
done on conversion, when creating a writable DWARF the existing logic
is kept, and source expects DW_FORM_string.