write/line: support any form for file source code (#786)
We can rely on the fact that if any source is set then the
corresponding string table will have a null byte in it,
and we can point to that for the empty string.

This is a bit of a hack, but it's workaround for an extension
to DWARF 5 only. DWARF 6 should have a better way of handling
this.