Reduce the size of `UnitSectionOffset`
Change `UnitSectionOffset` from an enum to a simple newtype, and store a
`SectionId` in `read::UnitHeader` instead.

`UnitSectionOffset` is intended for code that may be handling either
`.debug_info` or `.debug_types`. However, you will usually have a
`UnitHeader` you can use to determine the section type.

This significantly reduces the memory usage in the conversion code,
where we need to store an offset for every entry in the section.

This required some changes to the `UnitSectionOffset` methods. To make
things more consistent, I also changed these methods to use `UnitHeader`
instead of `Unit`, and added a `Deref` impl to `Unit`.