Delete `AddressOrOffsetPair` for range and location lists
`AddressOrOffsetPair` was a hack to handle the way we always used
`Reader::read_address` for ranges in DWARF version <= 4. We were doing
this because we needed to use `read_address` for the case where the
range is setting a base address.

A better way to handle this is to only use `read_address` if the
first value is !0, signifying that the next value is a base address.

As part of this change, we no longer support writing `StartEnd`
or `StartLength` for DWARF version <= 4. Our previous support for
this wasn't correct, because consumers would still add the base
address.