Add `read::EntriesRaw::read_attribute_inline`
Also change `read::EntriesRaw::read_attribute` to never inline.
The goal is to let users decide if they want performance or code size,
while at the same time making benchmarks more reliable, instead of
depending on whether the compiler decided to inline or not.

The change to always inline the internal `parse_attribute` means that
the `EntriesCursor` and `EntriesTree` benchmarks are improved.