Store attributes in a `Vec` in `read::DebuggingInformationEntry`
This simplifies the interface by removing a lifetime, and allows
infallible iteration of the attributes. `AttrsIter` is no longer
required.

Note that this hurts performance. We'll do further changes to improve
that, but the best performance will still come from using `EntriesRaw`
instead.

However, performance improves when repeating iteration of the
attributes since they are already parsed. This can happen
when calling methods such as `DebuggingInformationEntry::attr`.