Implement Iterator and remove use of FallibleIterator
The Rust ecosystem in general hasn't taken up use of `FallibleIterator`,
preferring instead to use `Iterator` with `Item=Result<_>`.

Our own uses aren't hard to convert to use `Iterator` with the std
library alone. There may be other crates that can help too, such as
itertools.

We still implement `FallibleIterator` for existing users that want it, but
don't enable it by default or document it.

Co-authored-by: Kyle Huey <khuey@pernos.co>